+++ plugin/src/main/java/com/atlassian/plugin/connect/plugin/capabilities/provider/AbstractContentMacroModuleProvider.java
@@ -55,43 +55,50 @@ public abstract class AbstractContentMacroModuleProvider<T extends BaseContentMa
         this.connectAddonI18nManager = connectAddonI18nManager;
     }
 
-    protected abstract ModuleDescriptor createMacroModuleDescriptor(ConnectAddonBean addon, Plugin theConnectPlugin, T macroBean);
+    protected abstract ModuleDescriptor createMacroModuleDescriptor(ConnectModuleProviderContext moduleProviderContext,
+                                                                    Plugin theConnectPlugin, T macroBean);
 
-    public List<ModuleDescriptor> provideModules(ConnectAddonBean addon, Plugin theConnectPlugin, String jsonFieldName, List<T> beans)
+    public List<ModuleDescriptor> provideModules(ConnectModuleProviderContext moduleProviderContext,
+                                                 Plugin theConnectPlugin, String jsonFieldName, List<T> beans)
     {
         List<ModuleDescriptor> moduleDescriptors = newArrayList();
-        MacroI18nBuilder i18nBuilder = new MacroI18nBuilder(addon.getKey());
+
+        final ConnectAddonBean connectAddonBean = moduleProviderContext.getConnectAddonBean();
+        MacroI18nBuilder i18nBuilder = new MacroI18nBuilder(connectAddonBean.getKey());
 
         for (T bean : beans)
         {
-            moduleDescriptors.addAll(createModuleDescriptors(addon, theConnectPlugin, bean));
-            i18nBuilder.add(bean, addon);
+            moduleDescriptors.addAll(createModuleDescriptors(moduleProviderContext, theConnectPlugin, bean));
+            i18nBuilder.add(bean, connectAddonBean);
         }
 
         try
         {
-            connectAddonI18nManager.add(addon.getKey(), i18nBuilder.getI18nProperties());
+            connectAddonI18nManager.add(connectAddonBean.getKey(), i18nBuilder.getI18nProperties());
         }
         catch (IOException e)
         {
-            log.error("Unable to register I18n properties for addon: " + addon.getKey(), e);
+            log.error("Unable to register I18n properties for addon: " + connectAddonBean.getKey(), e);
         }
 
         return moduleDescriptors;
     }
 
-    protected List<ModuleDescriptor> createModuleDescriptors(ConnectAddonBean addon, Plugin theConnectPlugin, T macroBean)
+    protected List<ModuleDescriptor> createModuleDescriptors(ConnectModuleProviderContext moduleProviderContext,
+                                                             Plugin theConnectPlugin, T macroBean)
     {
         List<ModuleDescriptor> descriptors = newArrayList();
 
+        final ConnectAddonBean addon = moduleProviderContext.getConnectAddonBean();
+
         // The actual Macro module descriptor
-        descriptors.add(createMacroModuleDescriptor(addon, theConnectPlugin, macroBean));
+        descriptors.add(createMacroModuleDescriptor(moduleProviderContext, theConnectPlugin, macroBean));
 
         // Add a web item if the Macro is featured
         if (macroBean.isFeatured())
         {
             WebItemModuleBean featuredWebItem = createFeaturedWebItem(addon, macroBean);
-            descriptors.add(webItemModuleDescriptorFactory.createModuleDescriptor(addon, theConnectPlugin, featuredWebItem));
+            descriptors.add(webItemModuleDescriptorFactory.createModuleDescriptor(moduleProviderContext, theConnectPlugin, featuredWebItem));
 
             // Add a featured icon web resource
             if (macroBean.hasIcon())