+++ plugin/src/main/java/com/atlassian/plugin/connect/plugin/capabilities/provider/ConfigurePageModuleProvider.java
@@ -5,7 +5,6 @@ import java.util.List;
 
 import com.atlassian.plugin.ModuleDescriptor;
 import com.atlassian.plugin.Plugin;
-import com.atlassian.plugin.connect.modules.beans.ConnectAddonBean;
 import com.atlassian.plugin.connect.modules.beans.ConnectPageModuleBean;
 import com.atlassian.plugin.connect.plugin.capabilities.descriptor.ConfigurePageModuleDescriptor;
 import com.atlassian.plugin.connect.plugin.capabilities.descriptor.WebItemModuleDescriptorFactory;
@@ -33,16 +32,17 @@ public class ConfigurePageModuleProvider extends AbstractAdminPageModuleProvider
     }
 
     @Override
-    public List<ModuleDescriptor> provideModules(ConnectAddonBean addon, Plugin theConnectPlugin, String jsonFieldName, List<ConnectPageModuleBean> beans)
+    public List<ModuleDescriptor> provideModules(ConnectModuleProviderContext moduleProviderContext, Plugin theConnectPlugin, String jsonFieldName, List<ConnectPageModuleBean> beans)
     {
-        super.provideModules(addon, theConnectPlugin, jsonFieldName, beans);
+        super.provideModules(moduleProviderContext, theConnectPlugin, jsonFieldName, beans);
 
         if(null != beans && !beans.isEmpty())
         {
             ConnectPageModuleBean configBean = beans.get(0);
 
             ModuleDescriptor descriptor = new ConfigurePageModuleDescriptor();
-            descriptor.init(theConnectPlugin, new DOMElement("connectConfigurePage").addAttribute("key", configBean.getKey(addon)));
+            descriptor.init(theConnectPlugin, new DOMElement("connectConfigurePage").addAttribute("key",
+                    configBean.getKey(moduleProviderContext.getConnectAddonBean())));
 
             return ImmutableList.of(descriptor);
         }