+++ plugin/src/main/java/com/atlassian/plugin/connect/plugin/rest/data/RestAddonType.java
@@ -1,0 +1,7 @@
+package com.atlassian.plugin.connect.plugin.rest.data;
+
+public enum RestAddonType
+{
+    XML,
+    JSON
+}