+++ plugin/src/main/resources/js/iframe/plugin/env.js
@@ -52,16 +52,6 @@ AP.define("env", ["_dollar", "_rpc"], function ($, rpc) {
         },
 
         /**
-        * fire an analytics event
-        *
-        * @param id  the event id.  Will be prepended with the prefix "p3.iframe."
-        * @param props the event properties
-        * @deprecated
-        */
-        fireEvent: function (id, props) {
-          console.log("AP.fireEvent deprecated; will be removed in future version");
-        },
-        /**
         * resize this iframe
         * @method
         * @param {String} width   the desired width