+++ plugin/src/test/java/com/atlassian/plugin/connect/test/plugin/capabilities/provider/SpaceToolsTabModuleProviderTest.java
@@ -10,6 +10,8 @@ import com.atlassian.plugin.connect.modules.beans.nested.XWorkInterceptorBean;
 import com.atlassian.plugin.connect.plugin.capabilities.ConvertToWiredTest;
 import com.atlassian.plugin.connect.plugin.capabilities.descriptor.WebItemModuleDescriptorFactory;
 import com.atlassian.plugin.connect.plugin.capabilities.descriptor.XWorkActionDescriptorFactory;
+import com.atlassian.plugin.connect.plugin.capabilities.provider.ConnectModuleProviderContext;
+import com.atlassian.plugin.connect.plugin.capabilities.provider.DefaultConnectModuleProviderContext;
 import com.atlassian.plugin.connect.plugin.capabilities.provider.SpaceToolsTabModuleProvider;
 import com.atlassian.plugin.connect.plugin.iframe.render.strategy.IFrameRenderStrategyBuilderFactory;
 import com.atlassian.plugin.connect.plugin.iframe.render.strategy.IFrameRenderStrategyRegistry;
@@ -54,6 +56,7 @@ public class SpaceToolsTabModuleProviderTest
     @Mock private IFrameRenderStrategyRegistry iFrameRenderStrategyRegistry;
     @Mock private IFrameRenderStrategyBuilderFactory iFrameRenderStrategyBuilderFactory;
 
+    private ConnectModuleProviderContext moduleProviderContext;
     private ConnectAddonBean addon;
     private SpaceToolsTabModuleProvider provider;
 
@@ -61,6 +64,7 @@ public class SpaceToolsTabModuleProviderTest
     public void setup()
     {
         this.addon = newConnectAddonBean().withKey("my-plugin").build();
+        this.moduleProviderContext = new DefaultConnectModuleProviderContext(addon);
         provider = new SpaceToolsTabModuleProvider(webItemModuleDescriptorFactory, xWorkActionDescriptorFactory,
                 productAccessor, iFrameRenderStrategyBuilderFactory, iFrameRenderStrategyRegistry);
         when(plugin.getKey()).thenReturn("my-plugin");
@@ -74,7 +78,7 @@ public class SpaceToolsTabModuleProviderTest
                 .withLocation("test-location")
                 .build();
 
-        provider.provideModules(addon, plugin, "spaceTools", ImmutableList.of(bean));
+        provider.provideModules(moduleProviderContext, plugin, "spaceTools", ImmutableList.of(bean));
 
         WebItemBeans webItems = captureWebItemBeans();
 
@@ -95,7 +99,7 @@ public class SpaceToolsTabModuleProviderTest
     @Test
     public void testWebItemUrl()
     {
-        provider.provideModules(addon, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
+        provider.provideModules(moduleProviderContext, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
 
         WebItemBeans webItems = captureWebItemBeans();
         String expectedUrl = "/plugins/atlassian-connect/my-plugin/test-module.action?key=${space.key}";
@@ -107,7 +111,7 @@ public class SpaceToolsTabModuleProviderTest
     public void testWebItemDefaultWeight()
     {
         when(productAccessor.getPreferredGeneralWeight()).thenReturn(666);
-        provider.provideModules(addon, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
+        provider.provideModules(moduleProviderContext, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
 
         WebItemBeans webItems = captureWebItemBeans();
         assertEquals(666, webItems.spaceTools.getWeight());
@@ -117,7 +121,7 @@ public class SpaceToolsTabModuleProviderTest
     @Test
     public void testSpaceToolsWebItemDefaultLocation()
     {
-        provider.provideModules(addon, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
+        provider.provideModules(moduleProviderContext, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
 
         WebItemBeans webItems = captureWebItemBeans();
         assertEquals(SpaceToolsTabModuleProvider.SPACE_TOOLS_SECTION + "/" + SpaceToolsTabModuleProvider.DEFAULT_LOCATION, webItems.spaceTools.getLocation());
@@ -127,7 +131,7 @@ public class SpaceToolsTabModuleProviderTest
     @Test
     public void testAction()
     {
-        provider.provideModules(addon, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
+        provider.provideModules(moduleProviderContext, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
         XWorkActionModuleBean actionModuleBean = captureActionBean();
 
         assertEquals("/plugins/atlassian-connect/my-plugin", actionModuleBean.getNamespace());
@@ -139,7 +143,7 @@ public class SpaceToolsTabModuleProviderTest
     @Ignore
     public void testActionContextData()
     {
-        provider.provideModules(addon, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
+        provider.provideModules(moduleProviderContext, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
         XWorkActionModuleBean actionModuleBean = captureActionBean();
 
         SpaceToolsTabContext context = (SpaceToolsTabContext) actionModuleBean.getParameters().get("context");
@@ -154,7 +158,7 @@ public class SpaceToolsTabModuleProviderTest
     @Test
     public void testActionInterceptor()
     {
-        provider.provideModules(addon, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
+        provider.provideModules(moduleProviderContext, plugin, "spaceTools", ImmutableList.of(DEFAULTS_BEAN));
         XWorkActionModuleBean actionModuleBean = captureActionBean();
 
         XWorkInterceptorBean interceptorBean = actionModuleBean.getInterceptorsBeans().get(0);
@@ -172,7 +176,7 @@ public class SpaceToolsTabModuleProviderTest
     private WebItemBeans captureWebItemBeans()
     {
         ArgumentCaptor<WebItemModuleBean> captor = ArgumentCaptor.forClass(WebItemModuleBean.class);
-        verify(webItemModuleDescriptorFactory, times(2)).createModuleDescriptor(addon, eq(plugin), captor.capture());
+        verify(webItemModuleDescriptorFactory, times(2)).createModuleDescriptor(moduleProviderContext, eq(plugin), captor.capture());
         List<WebItemModuleBean> beans = captor.getAllValues();
         return new WebItemBeans(beans.get(0), beans.get(1));
     }