+++ docs/public/resources/release-notes.md
@@ -1,7 +1,17 @@
 
 # Release Notes
 
-## [1.1.0-beta.4](../release-notes/1-1-0.html)
+## [1.1.0-rc.1](../release-notes/1-1-0.html)
+* __If you read one thing, read about [breaking changes](../release-notes/1-1-0.html#breaking-changes)__
+* JavaScript API: removed deprecated function `AP.fireEvent()`
+* New feature: condition for checking dark features
+* Fixed: web-sections which rely on add-on-provided web-items fail to register
+* Valid module keys are no longer modified
+* Whitelisted Confluence Questions [context params](../concepts/context-parameters.html) for webitems and webpanels
+* Docs: Fix broken links to webhooks module page
+* Docs: web panels are not to be used for dialog content
+
+## [1.1.0-beta.5](../release-notes/1-1-0.html)
 * __If you read one thing, read about [breaking changes](../release-notes/1-1-0.html#breaking-changes)__
 * XML descriptor servlet paths redirect to new JSON descriptor paths
 * Manage cookies through the [javascript cookie api](http://localhost:9000/javascript/module-cookie.html)
@@ -17,14 +27,14 @@
 * Improved documentation for workflow post functions
 * Remote web panels fixed in JIRA Agile
 
-## [1.0.1](../release-notes/1-0-1.html)
-* Allow POST method for "screens/addToDefault/{fieldId}"
-* UPM auto-update fails to upgrade add-on
-
 ## [1.0.2](../release-notes/1-0-2.html)
 * Fixed workflow post functions
 * Support for hidden macros in the macro browser
 * Removing content from Web-panel no longer leaves grey bar in place
+
+## [1.0.1](../release-notes/1-0-1.html)
+* Allow POST method for "screens/addToDefault/{fieldId}"
+* UPM auto-update fails to upgrade add-on
 
 ## [1.0.0](../release-notes/1-0.html)
 * Installing an add-on into OnDemand will not work unless the base url starts with https